coMMit before entering a new database

+++ domain/src/main/java/com/BySandS/testsandquizes/domain/tests/models/param/GetQuestionListParam.kt
@@ -2,5 +2,5 @@ package com.BySandS.testsandquizes.domain.tests.models.param
 
 class GetQuestionListParam (
     val difficultyId: Long,
-    val quantityOfQuestions: Int
+    val quantityOfQuestions: Long
 )
